# Avoid memory allocations when comparing chunk keys