@tomwilkie 2017-12-22
Is this safe?  There are at least two different version of chunk keys, one from before we added checksums.