@bboreham 2017-12-22
> Is this safe? There are at least two different version of chunk keys, one from before we added checksums.

There are exactly two versions in `Chunk.ExternalKey()`, differentiated by whether the checksum is set.

Since we compare everything else ahead of the checksum, we are left to contemplate the problem of two chunks which are identical except one was stored with and one stored without the checksum.  And the checksum is exactly zero.  The new code would consider those chunks equal, and the old would consider them different.  It would be easy to re-instate this behaviour, but I'd like to hear the justification.

I remarked in the commit message on the more general theme of changing the sort order: it was fairly arbitrary before, and it is different now.  But as long as the order is self-consistent and matches the equality test used for de-duping I think it is fit for purpose.