@bboreham 2018-01-03
I updated `lessByKey` roughly as suggested by @rade.

Any more thoughts?