@bboreham 2018-01-04
Rebased and extended to `Chunk.Decode()`